# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rahul Raj — Portfolio & Systems Architecture API",
    "version": "1.0.0",
    "description": "Production-grade RESTful API documentation for Rahul Raj's modern portfolio platform, built with **Next.js 16**, **TypeScript**, **MongoDB Atlas**, **Cloudinary**, and **AI LLM Orchestration**.\n\n### Key Architectural Pillars\n- **Public Gateway**: Ultra-fast endpoints for portfolio contents, live ATS resume downloads, real-time presence heartbeats, and AI assistant interaction.\n- **Admin Control Studio & Hardware Biometrics**: Role-Based Access Control (RBAC) secured via WebAuthn/FIDO2 hardware biometric passkeys (Fingerprint, TouchID, Windows Hello, FaceID), 4-digit security PIN, master passcode, cryptographic HMAC tokens, or Email OTP resets.\n- **Resilience Engine**: Zero-crash persistence architecture backed by MongoDB Atlas with in-memory caching and real-time health telemetry probes.\n- **Security & Anti-Abuse**: IP-based sliding window rate limiters, anti-bot honeypot traps, and encrypted session cookies.",
    "contact": {
      "name": "Rahul Raj Modi",
      "email": "rahulraj21480@gmail.com",
      "url": "https://github.com/Rahul-2148"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Environment / Origin Server"
    },
    {
      "url": "http://localhost:3000",
...
    }
  ],
...
  "paths": {
...
    "/api/admin/auth": {
      "post": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Authenticate via biometric passkey, 4-digit PIN, or master passcode",
        "description": "Multi-tier authentication endpoint supporting hardware biometric passkeys (WebAuthn / FIDO2 Fingerprint, TouchID, Windows Hello), 4-digit security PIN, master passcode, setting HTTP-only HMAC session cookies, requesting 6-digit Email OTP, resetting passcode via OTP, or logging out.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "login",
                      "logout",
                      "change_passcode",
                      "send_email_otp",
                      "verify_email_otp"
                    ],
                    "default": "login"
                  },
                  "passcode": {
                    "type": "string",
                    "description": "Master passcode for login"
                  },
                  "currentPasscode": {
                    "type": "string",
                    "description": "Used during change_passcode"
                  },
                  "newPasscode": {
                    "type": "string",
                    "description": "Used during change_passcode or verify_email_otp"
                  },
                  "otp": {
                    "type": "string",
                    "description": "6-digit verification code"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication action successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Authentication successful. Welcome to Admin Studio."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or expired OTP"
          }
        }
      },
      "get": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Check admin authentication status",
        "description": "Inspects active session cookie or header passcode and returns authentication state.",
        "responses": {
          "200": {
            "description": "Authentication state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authenticated": {
                      "type": "boolean",
                      "example": true
                    },
                    "adminEmail": {
                      "type": "string",
                      "example": "rahulraj21480@gmail.com"
                    },
                    "hasCustomPasscode": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/visitor/auth": {
      "post": {
        "tags": [
          "Visitor & Recruiter Pass"
        ],
        "summary": "Issue Recruiter or Visitor Pass",
        "description": "Verifies and records recruiter/visitor credentials, associates them with the active telemetry session, and upgrades anonymous analytics to a verified recruiter lead.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Sarah Connor"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "sarah@techcorp.io"
                  },
                  "company": {
                    "type": "string",
                    "example": "TechCorp Systems"
                  },
                  "role": {
                    "type": "string",
                    "example": "Lead Technical Recruiter"
                  },
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "hiring",
                      "mentorship",
                      "collaboration",
                      "casual"
                    ],
                    "example": "hiring"
                  },
                  "notes": {
                    "type": "string",
                    "example": "Interested in distributed systems background"
                  },
                  "visitorId": {
                    "type": "string",
                    "example": "vis_1711456789_abc123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visitor pass successfully generated and session upgraded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "Sarah Connor"
                        },
                        "email": {
                          "type": "string",
                          "example": "sarah@techcorp.io"
                        },
                        "company": {
                          "type": "string",
                          "example": "TechCorp Systems"
                        },
                        "role": {
                          "type": "string",
                          "example": "Lead Technical Recruiter"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Welcome Sarah Connor! Recruiter Pass generated successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error — missing required fields"
          },
          "500": {
            "description": "Server error processing pass registration"
          }
        }
      }
    },
    "/api/admin/cloudinary/delete": {
      "post": {
        "tags": [
          "Admin Settings & Media"
        ],
        "summary": "Delete media assets from Cloudinary CDN",
        "description": "Deletes a single asset or performs bulk deletion of multiple media assets by public ID. Logs administrative audit trail upon success.",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "publicId": {
                    "type": "string",
                    "description": "Single asset public ID to delete",
                    "example": "portfolio/projects/demo_image"
                  },
                  "publicIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Bulk array of public IDs to delete"
                  },
                  "resourceType": {
                    "type": "string",
                    "enum": [
                      "image",
                      "raw",
                      "video"
                    ],
                    "default": "image",
                    "example": "image"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asset(s) successfully deleted from Cloudinary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Deleted 3 assets."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing publicId or publicIds in request body"
          },
          "401": {
            "description": "Unauthorized — valid admin session or passcode required"
          },
          "500": {
            "description": "Error communicating with Cloudinary API"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminCookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "portfolio_admin_token",
        "description": "Encrypted HTTP-only admin session token generated during login"
      },
      "AdminPasscodeAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-passcode",
        "description": "Secret master passcode configured via ADMIN_PASSCODE environment variable"
      }
    },
...
  }
}